Extract catalog routes and gallery media out of submissions.ts
Phase 3 Wave B: pulls the read-only public catalog (loadCatalog,
getCatalogEntries, storeCatalogEntries, deattributeDeletedOwners),
store-published game lookups, gallery media serving, and the semantic
catalog search into apps/api/src/catalog/catalog-routes.ts and
catalog-search-routes.ts, registered from within registerSubmissionRoutes.
Also lifts the generic isRateLimited/sendMedia/parseBytesRange helpers
into apps/api/src/platform/ip-rate-limit.ts and media-response.ts so
both submissions.ts and the new catalog module share them.

submissions.ts drops from 6427 to 5791 lines with no behavior change:
/api/games/:slug's draft-preview fallback (canPlayDraft/replyWithDraft)
stays behind since it pulls in the whole draft-preview subsystem, and
the gallery-media route keeps sharing its per-IP rate-limit bucket with
the build-screenshot route exactly as before.
