+++ apps/api/src/platform/ip-rate-limit.ts
@@ -1,0 +1,18 @@
+// Sliding-window rate limiter keyed by client IP.
+export function isRateLimited(
+  buckets: Map<string, number[]>,
+  ip: string,
+  currentTime: number,
+  maxRequests: number,
+  windowMs: number,
+): boolean {
+  const requests = (buckets.get(ip) ?? []).filter((timestamp) => currentTime - timestamp < windowMs);
+  if (requests.length >= maxRequests) {
+    buckets.set(ip, requests);
+    return true;
+  }
+
+  requests.push(currentTime);
+  buckets.set(ip, requests);
+  return false;
+}
